URL check and contents check added

+++ Web Scraper/task/scraper.py
@@ -1,6 +1,7 @@
 import os
 import shutil
 import string
+
 import requests
 from bs4 import BeautifulSoup
 
@@ -12,41 +13,56 @@ list_of_articles = list()
 def extract_links(pages_number, articles_type):
     for page_number in range(pages_number):
         request = requests.get(WEBSITE_URL + f'&page={page_number + 1}')
-        soup = BeautifulSoup(request.content, 'html.parser')
 
-        if os.access(f'./Page_{page_number + 1}', os.F_OK):
-            shutil.rmtree(f'./Page_{page_number + 1}')
-        os.mkdir(f'./Page_{page_number + 1}')
+        if request.status_code == 200:
+            soup = BeautifulSoup(request.content, 'html.parser')
 
-        articles = soup.find_all('article')
-        for article in articles:
-            article_type = article.find('span', attrs={'c-meta__item'}).text
-            if article_type == articles_type:
-                list_of_articles.append(article.find('a').get('href'))
-        search_for_details(page_number)
-        list_of_articles.clear()
+            if os.access(f'./Page_{page_number + 1}', os.F_OK):
+                print('Ok')
+                shutil.rmtree(f'./Page_{page_number + 1}')
+            os.mkdir(f'./Page_{page_number + 1}')
+
+            articles = soup.find_all('article')
+            for article in articles:
+                article_type = article.find('span', attrs={'c-meta__item'}).text
+                if article_type == articles_type:
+                    list_of_articles.append(article.find('a').get('href'))
+
+            search_for_details(page_number)
+            list_of_articles.clear()
+        else:
+            print('URL is incorrect!')
+
+    print('Saved all articles.')
 
 
 def search_for_details(page_number):
     for article in list_of_articles:
         request = requests.get(DOMAIN + article)
-        soup = BeautifulSoup(request.content, 'html.parser')
 
-        article_title = (soup.find('title')
-                         .text
-                         .translate(str.maketrans('', '', string.punctuation))
-                         .replace(' ', '_')
-                         .strip())
+        if request.status_code == 200:
+            soup = BeautifulSoup(request.content, 'html.parser')
 
-        article_content = (soup.find('p', {'class': 'article__teaser'})
-                           .text
-                           .strip())
+            article_title = (soup.find('title')
+                             .text
+                             .translate(str.maketrans('', '', string.punctuation))
+                             .replace(' ', '_')
+                             .strip())
 
-        file = open(f'./Page_{page_number + 1}/' + article_title + '.txt',
-                    mode='w',
-                    encoding='UTF-8')
-        file.write(article_content)
-        file.close()
+            if soup.find('p', {'class': 'article__teaser'}) is None:
+                article_content = ''
+            else:
+                article_content = (soup.find('p', {'class': 'article__teaser'})
+                                   .text
+                                   .strip())
+
+            file = open(f'./Page_{page_number + 1}/' + article_title + '.txt',
+                        mode='w',
+                        encoding='UTF-8')
+            file.write(article_content)
+            file.close()
+        else:
+            print('URL is incorrect!')
 
 
 number_of_pages = int(input())
